+++ apps/meteor/tests/e2e/create-direct.spec.ts
@@ -14,7 +14,7 @@ test.describe.serial('channel-direct-message', () => {
 	});
 
 	test('expect create a direct room', async ({ page }) => {
-		await poHomeChannel.sidenav.openNewByLabel('Direct Messages');
+		await poHomeChannel.sidenav.openNewByLabel('Direct messages');
 
 		await poHomeChannel.sidenav.inputDirectUsername.click();
 		await page.keyboard.type('rocket.cat');